services/graphrag/requirements.txt
fastapi
uvicorn
falkordb
pydantic
python-dotenv
langchain-core
langfuse
llama-index-core
mcp
psycopg2-binary
pgvector
pydantic-settings
llama-index-embeddings-huggingface
llama-index-llms-groq
llama-index-llms-openai
torch
transformers
tiktoken
langfuse
openinference-instrumentation-llama-index